+++ docs/history.rst
@@ -27,3 +27,8 @@ New in version 0.2.0:
 - Extended, improved and customized documentation website (generated with sphinx).
 - Extended and improved Jupyter Notebook for documentation.
 - Improved visualisation for *get_correlation*: a prediction band based on the obtained correlation is now included in the produced scatter plot.
+
+Known bugs (see also (open issues on Github)[https://github.com/UGentBiomath/wwdata/issues]):
+
+- The ``fill_missing_daybefore`` function uses the data of the day before the given ``arange`` period to fill gaps in the data. This data is not necessarily the data from the actual day before the period where data is filled.
+-